Put HomeScreen function into separate Screens folder
