+++ App.js
@@ -6,51 +6,7 @@ import LoginScreen from './forms/login.js';
 import RegisterScreen from './forms/register.js';
 import Swiper from 'react-native-deck-swiper';
 import { Card } from './components/Cards.js';
-
-function HomeScreen({ navigation }) {
-  const [isLoading, setLoading] = useState(true);
-  const [data, setData] = useState([]);
-
-  useEffect(() => {
-    fetch('https://api.themoviedb.org/3/movie/popular?api_key=156f6cfa04dae615351cd9878f39b732&language=en-US&page=1')
-      .then((response) => response.json())
-      .then((json) => setData(parseMovies(json.results)))
-      .catch((error) => console.error(error))
-      .finally(() => setLoading(false));
-  }, []);
-  console.log(data);
-  return (
-    <SafeAreaView style={styles.container}>
-      {isLoading ? <ActivityIndicator/> : (
-        <Swiper
-            cards={data}
-            renderCard={Card}
-            infinite // keep looping cards infinitely
-            verticalSwipe={false}
-            backgroundColor="white"
-            cardHorizontalMargin={0}
-            stackSize={2} // number of cards shown in background
-            />
-      )}
-    </SafeAreaView>    
-  );
-}
-
-function parseMovies(movieArray) {
-  var parsedMovies = [];
-  var i;
-  var imgurl= "https://image.tmdb.org/t/p/original";
-  for (i = 0; i < movieArray.length; i++) {
-    parsedMovies[i] =
-    {
-        pic: {uri: imgurl.concat(movieArray[i].poster_path)},
-        title: movieArray[i].title,
-        caption: movieArray[i].overview,
-    }
-  }
-  return parsedMovies
-}
-
+import HomeScreen from './screens/HomeScreen.js';
 
 function LogoTitle() {
   return (
@@ -60,23 +16,6 @@ function LogoTitle() {
     />
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    justifyContent: 'center',
-    alignItems: "center",
-  },
-  logout_button: {
-    width: "80%",
-    borderRadius: 25,
-    alignItems: "center",
-    backgroundColor: "#99ccff",
-    marginTop: 100,
-    padding: 10,
-  },
-})
 
 const Stack = createStackNavigator();
